Add cleaUp.gs from my google app scripts
I use this to keep my gmail inbox cleaned out of crap that accumulates from automatic mailing lists. I've had it for years, need to tinker with the permissions now and again, probably better to upload it. It's pretty simple in that it just goes though all the "delete me" labels and sublabels and chucks anything older than 10 days in the trash. It's up to the user to set up filters which handle incoming mail appropriately.